Move ReaderCommonOptions class to Common folder (#343)
