+++ src/Spout/Reader/ODS/ReaderOptions.php
@@ -2,15 +2,13 @@
 
 namespace Box\Spout\Reader\ODS;
 
-use Box\Spout\Reader\ReaderOptionsCommon;
-
 /**
  * Class ReaderOptions
  * This class is used to customize the reader's behavior
  *
  * @package Box\Spout\Reader\ODS
  */
-class ReaderOptions extends ReaderOptionsCommon
+class ReaderOptions extends \Box\Spout\Reader\Common\ReaderOptions
 {
     // No extra options
 }
